requirements.txt
asgiref
certifi
contourpy
crispy-bootstrap5
cycler
distlib
Django
django-crispy-forms
django-environ
django-filter
djangorestframework
filelock
fonttools
kiwisolver
Markdown
matplotlib
MouseInfo
numpy
packaging
pillow
pipenv
platformdirs
PyAutoGUI
PyGetWindow
PyMsgBox
pyparsing
pyperclip
PyRect
PyScreeze
python-dateutil
pytweening
setuptools
six
sqlparse
tzdata
virtualenv